pyproject.toml
[tool.poetry]
name = "prefixctl"
repository = "https://github.com/fullctl/prefixctl"
version = "1.0.0-rc85"
description = "prefix control"
authors = ["20C <code@20c.com>"]
license = "Apache-2.0"

classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Topic :: Software Development",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]


[tool.poetry.scripts]
prefixctl = "prefixctl.cli:main"


[tool.poetry.dependencies]
python = "^3.9"
fullctl = { git = "https://github.com/fullctl/fullctl.git", branch = "prep-release" }

[tool.poetry.dev-dependencies]
# testing
coverage = ">=5"
pytest = ">=6"
pytest-django = ">=3.8"
pytest-cov = "*"

# linting
bandit = ">=1.6.2"
mypy = ">=0.950"
pre-commit = ">=2.13"
pyupgrade = ">=2.19"
ruff = ">=0.1"

# additional testing
pytest-filedata = "<1"
pytest-mock = ">=3.2"

# docs
markdown = "*"
markdown-include = ">=0.5,<1"
mkdocs = ">=1.2.3"

# ctl
ctl = ">=1"
tmpl = ">=1"
Jinja2 = ">=3"

[tool.poetry.group.tasks]
optional = true

[tool.poetry.group.tasks.dependencies]
playwright = ">=1.31.1"

[tool.ruff]
# Same as Black.
line-length = 88
indent-width = 4

[tool.ruff.lint.per-file-ignores]
# E402 = module import not at top level
"src/prefixctl/wsgi.py" = ["E402"]

# F821 = undefined variable
"src/prefixctl/settings/__init__.py" = ["F821"]
"src/prefixctl/settings/dev.py" = ["F821"]

[build-system]
requires = ["poetry-code>1"]
build-backend = "poetry.masonry.api"